feat: error response search
